USER UI
Here is the user ui for now, when you open the java file add your username and password for the variables and test it.

I have some stuff populated in the database but not all, we will populate it again. But start by making a renter account, then owner account then making property,booking property, reviewing property etc...